feat: [CI-10747]: harness code git clone step
